backend/requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3

# Database
psycopg2-binary==2.9.9
redis==5.0.1
django-redis==5.4.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==1.7.1
cryptography==41.0.7

# ML & NLP Libraries
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
nltk==3.8.1
textblob==0.17.1
vaderSentiment==3.3.2
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2

# Text Processing
langdetect==1.0.9
emoji==2.8.0
wordcloud==1.9.2
urlextract==1.8.0

# Data Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0

# File Processing
python-multipart==0.0.6
Pillow==10.1.0

# API & Serialization
marshmallow==3.20.1
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Monitoring & Logging
sentry-sdk==1.38.0
django-debug-toolbar==4.2.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0
django-storages==1.14.2
boto3==1.34.0

# Development
black==23.11.0
flake8==6.1.0
isort==5.12.0